wl: Move declaration of CogWlPlatform to a header file
